backend/requirements.txt
# Backend requirements
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-dotenv==1.0.0
openai==1.3.0
numpy==1.25.2
scikit-learn==1.3.1
pandas==2.1.1
pymongo==4.5.0
motor==3.3.1
python-jose==3.3.0
python-multipart==0.0.6
bcrypt==4.0.1
passlib==1.7.4